Move refactored message to __init__ and fix cyclic imports
